requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-cleanup==5.2.0
django-environ==0.4.5
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
